Add a `PrivateSigningKey::inner_key` trait method.
Goal is to make is easier to reuse the the signing keys for other
libraries.

Removes the private `EdDsaSigningKey` enum which aims to adds support
more curves than that of `Ed25517`, even though it's the only one
in use. `CoreEdDsaPrivateSigningKey` could instead be made generic to
support this feature, if it is ever expanded upon.